CXP: revamp steps in the wizard
