+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/AppWizard/AuthenticationModal.tsx
@@ -41,13 +41,19 @@ export const AuthenticationModal: FC<Props> = ({clientId}) => {
 
     useEffect(() => {
         fetchWizardData().then(wizardData => {
+            const hasAlreadyConsented =
+                !!wizardData.oldAuthenticationScopes && wizardData.oldAuthenticationScopes.length > 0;
+
             setWizardData(wizardData);
+            setScopesConsent(hasAlreadyConsented);
         });
     }, [fetchWizardData]);
 
     if (!wizardData) {
         return null;
     }
+
+    const hasAlreadyConsented = !!wizardData.oldAuthenticationScopes && wizardData.oldAuthenticationScopes.length > 0;
 
     return (
         <WizardModal
@@ -59,7 +65,7 @@ export const AuthenticationModal: FC<Props> = ({clientId}) => {
             steps={[
                 {
                     name: 'authentication',
-                    action: 'confirm',
+                    requires_explicit_approval: true,
                 },
             ]}
         >
@@ -69,6 +75,7 @@ export const AuthenticationModal: FC<Props> = ({clientId}) => {
                     scopes={wizardData.authenticationScopes}
                     oldScopes={wizardData.oldAuthenticationScopes}
                     appUrl={wizardData.appUrl}
+                    skipConsent={hasAlreadyConsented}
                     scopesConsentGiven={scopesConsentGiven}
                     setScopesConsent={setScopesConsent}
                 />
