+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/AppWizard/WizardModal.tsx
@@ -3,6 +3,7 @@ import React, {FC} from 'react';
 import styled from 'styled-components';
 import {useTranslate} from '../../../shared/translate';
 import {useStepProgress} from './useStepProgress';
+import getStepConfirmationLabel from './getStepConfirmationLabel';
 
 const Content = styled.div`
     display: grid;
@@ -28,9 +29,9 @@ const ProgressIndicatorContainer = styled(ProgressIndicator)`
     bottom: 20px;
 `;
 
-type Step = {
+export type Step = {
     name: 'authentication' | 'authorizations' | 'permissions' | 'summary';
-    action: 'next' | 'allow_and_next' | 'confirm' | 'allow_and_finish';
+    requires_explicit_approval: boolean;
 };
 
 type Props = {
@@ -54,7 +55,9 @@ export const WizardModal: FC<Props> = ({
 }) => {
     const translate = useTranslate();
 
-    const {current, isFirst, isLast: isLast, next, previous} = useStepProgress(steps);
+    const {current, isFirst, isLast, next, previous} = useStepProgress(steps);
+
+    const confirmLabel = getStepConfirmationLabel(current, isFirst, isLast);
 
     const isSingleStepWizard = steps.length === 1;
 
@@ -71,7 +74,7 @@ export const WizardModal: FC<Props> = ({
                         </Button>
                     )}
                     <Button onClick={isLast ? onConfirm : next} disabled={current.name === maxAllowedStep}>
-                        {translate(`akeneo_connectivity.connection.connect.apps.wizard.action.${current.action}`)}
+                        {translate(confirmLabel)}
                     </Button>
                 </Modal.TopRightButtons>
             )}
@@ -99,7 +102,7 @@ export const WizardModal: FC<Props> = ({
                         {translate('akeneo_connectivity.connection.connect.apps.wizard.action.cancel')}
                     </Button>
                     <Button onClick={onConfirm} disabled={current.name === maxAllowedStep}>
-                        {translate('akeneo_connectivity.connection.connect.apps.wizard.action.confirm')}
+                        {translate(confirmLabel)}
                     </Button>
                 </Modal.BottomButtons>
             )}
