+++ src/Akeneo/Connectivity/Connection/front/src/connect/components/AppWizard/steps/Authentication/Authentication.tsx
@@ -49,6 +49,7 @@ type Props = {
     appUrl: string | null;
     scopes: Array<'email' | 'profile'>;
     oldScopes?: Array<'email' | 'profile'> | null;
+    skipConsent: boolean;
     scopesConsentGiven: boolean;
     setScopesConsent: (newValue: boolean) => void;
 };
@@ -58,6 +59,7 @@ export const Authentication: FC<Props> = ({
     appUrl,
     scopes,
     oldScopes,
+    skipConsent,
     scopesConsentGiven,
     setScopesConsent,
 }) => {
@@ -89,7 +91,9 @@ export const Authentication: FC<Props> = ({
                     <ConsentList scopes={oldScopes} highlightMode={'old'} />
                 </>
             )}
-            <ConsentCheckbox isChecked={scopesConsentGiven} onChange={setScopesConsent} appUrl={appUrl} />
+            {!skipConsent && (
+                <ConsentCheckbox isChecked={scopesConsentGiven} onChange={setScopesConsent} appUrl={appUrl} />
+            )}
         </InfoContainer>
     );
 };
