+++ src/Akeneo/Connectivity/Connection/front/tests/src/connect/components/AppWizard/steps/Authentication/Authentication.test.tsx
@@ -29,6 +29,7 @@ test('it renders correctly for first connection', async () => {
             scopes={['email']}
             oldScopes={null}
             appUrl={null}
+            skipConsent={false}
             scopesConsentGiven={false}
             setScopesConsent={() => null}
         />
@@ -57,6 +58,7 @@ test('it renders correctly with new scopes required when not already had old sco
             scopes={['email']}
             oldScopes={[]}
             appUrl={null}
+            skipConsent={false}
             scopesConsentGiven={false}
             setScopesConsent={() => null}
         />
@@ -85,6 +87,7 @@ test('it renders correctly with new scopes required when already accepted old sc
             scopes={['email']}
             oldScopes={['profile']}
             appUrl={null}
+            skipConsent={false}
             scopesConsentGiven={false}
             setScopesConsent={() => null}
         />
